Map namespaces in kcp to unique names in pclusters
A namespace gets translated to a template that includes the kcp
workspace name.

Example: syncer-test -> kcp--admin--syncer-test

This change required that the syncer start processing namespaces (the
only non-namespaced resource currently processed by the syncer).  The
`process` method was modified to handle both namespaces and namespaced
resources.

This includes an e2e test that creates a namespace and a custom resource
in that namespace and validates that it gets sync'd to a downstream
cluster using the translated namespace.

There is (at least on my local box) a mysterious ~60s delay in the
namespace scheduler in reconciling and applying the correct label to the
custom resource in the e2e test.

Fixes #283

Signed-off-by: Kyle Lape <klape@redhat.com>