Add kcp-config argument wrapper
usage: kcp-config [config file] [cmd] [args...]

This allows kcp to be configured using a yaml file.  The yaml document
needs to be a flat map (i.e. no hierarchy), and each key/value pair will
be added as a command-line argument to kcp.

Example config.yaml:

    enable-controllers: true
    listen: localhost:6443

with this invocation:

    kcp-config config.yaml kcp start

will translate to:

    kcp start --enable-controllers=true --listen=localhost:6443